Re-base collectors' drawing half onto cellpy.plotting.
Move multi-cell plotters into collected_plot with layout/kind options so
BatchCollector.render delegates into the shared prepare→spec→render home
and collectors no longer own figure drawing.
