+++ .issueflows/03-solved-issues/issue657_status.md
@@ -1,0 +1,15 @@
+# Issue #657 — Status
+
+- [x] Done
+
+## What's done
+
+- Re-based collectors drawing onto `cellpy.plotting.collected_plot` (`layout=` / `kind=`).
+- Moved plotter bodies out of `collectors.py`; `BatchCollector.render` / `plot` delegate.
+- `FrameContext` + design note `plotting-collected.md`.
+- Collector figure oracle (`tests/data/collector_figure_specs.json`) + tests.
+- Focused suite green (collectors / plotting_package / figure_specs).
+
+## Remaining work
+
+- None.
