+++ tests/test_collectors.py
@@ -11,9 +11,17 @@ data collection and the plot) so a broken column-name migration or a plotter
 regression fails loudly. They require the plotting extras; the `full` CI job
 installs them (`uv sync --extra batch`), and they skip cleanly when the extras
 or the batch testdata are absent.
+
+Collector figure structure is snapshotted in
+``tests/data/collector_figure_specs.json`` (#657) — regenerate with::
+
+    MPLBACKEND=Agg uv run python -c "from tests.test_collectors import write_collector_figure_specs; write_collector_figure_specs()"
 """
 
 from __future__ import annotations
+
+import json
+from pathlib import Path
 
 import pytest
 
@@ -23,13 +31,19 @@ from tests.test_batch import (  # noqa: F401  (imported for fixture resolution)
     clean_dir,
     populated_batch,
 )
+from tests.figure_spec_support import describe_figure
 
 plotly = pytest.importorskip("plotly", reason="plotting extras (batch) not installed")
 
+from cellpy.utils import collectors as collectors_mod  # noqa: E402
 from cellpy.utils.collectors import (  # noqa: E402
     BatchCyclesCollector,
     BatchICACollector,
     BatchSummaryCollector,
+)
+
+COLLECTOR_SNAPSHOT_PATH = (
+    Path(__file__).resolve().parent / "data" / "collector_figure_specs.json"
 )
 
 
@@ -103,3 +117,99 @@ def test_select_direction_handles_both_encodings():
 
     without = pd.DataFrame({"v": [1, 2]})
     assert len(_select_direction(without, "charge")) == 2
+
+
+def test_drawing_bodies_live_in_plotting_not_collectors():
+    """Collectors no longer define sequence/summary/cycles/ica/spread plotters (#657)."""
+    for name in (
+        "sequence_plotter",
+        "summary_plotter",
+        "cycles_plotter",
+        "ica_plotter",
+        "spread_plot",
+        "_cycles_plotter",
+    ):
+        assert not hasattr(collectors_mod, name), name
+
+
+def test_batch_collector_plot_aliases_render(populated_batch):
+    """``BatchCollector.plot`` is a thin alias of ``render`` (#657)."""
+    collector = BatchSummaryCollector(populated_batch, autorun=False)
+    collector.collect()
+    collector.plot()
+    assert collector.figure is not None
+
+
+def _collector_figure_menu(populated_batch) -> dict:
+    """Minimum collector column for the #657 oracle."""
+    from cellpy.plotting import collected_plot
+
+    summary = BatchSummaryCollector(populated_batch, autorun=False)
+    summary.collect()
+    cycles = BatchCyclesCollector(populated_batch, autorun=False)
+    cycles.collect()
+    ica_film = BatchICACollector(populated_batch, plot_type="film", autorun=False)
+    ica_film.collect()
+    # spread_plot needs group_it so the frame carries mean/std columns
+    summary_spread = BatchSummaryCollector(
+        populated_batch, group_it=True, spread=True, autorun=False
+    )
+    summary_spread.collect()
+
+    figures = {
+        "collected_summary[plotly]": describe_figure(
+            collected_plot(summary.data, family_kind="summary", backend="plotly")
+        ),
+        "collected_cycles_per_cell[plotly]": describe_figure(
+            collected_plot(
+                cycles.data,
+                family_kind="cycles",
+                layout="per_cell",
+                backend="plotly",
+            )
+        ),
+        "collected_ica_film[plotly]": describe_figure(
+            collected_plot(
+                ica_film.data,
+                family_kind="ica",
+                kind="film",
+                backend="plotly",
+            )
+        ),
+        "collected_summary_spread[plotly]": describe_figure(
+            collected_plot(
+                summary_spread.data,
+                family_kind="summary",
+                kind="spread",
+                backend="plotly",
+            )
+        ),
+    }
+    return {"figures": figures}
+
+
+def write_collector_figure_specs(populated_batch=None) -> Path:
+    """Regenerate ``collector_figure_specs.json`` (dev helper / snapshot regen)."""
+    if populated_batch is None:
+        raise RuntimeError("pass a populated_batch when calling from tests")
+    specs = _collector_figure_menu(populated_batch)
+    COLLECTOR_SNAPSHOT_PATH.parent.mkdir(parents=True, exist_ok=True)
+    COLLECTOR_SNAPSHOT_PATH.write_text(
+        json.dumps(specs, indent=2, sort_keys=True) + "\n", encoding="utf-8"
+    )
+    return COLLECTOR_SNAPSHOT_PATH
+
+
+@pytest.mark.essential
+def test_collector_figure_structure_matches_snapshot(populated_batch):
+    """Collector layouts are part of the plotting contract (#657)."""
+    if not COLLECTOR_SNAPSHOT_PATH.is_file():
+        pytest.skip(f"missing snapshot {COLLECTOR_SNAPSHOT_PATH}")
+    expected = json.loads(COLLECTOR_SNAPSHOT_PATH.read_text(encoding="utf-8"))["figures"]
+    actual = _collector_figure_menu(populated_batch)["figures"]
+    assert set(actual) == set(expected)
+    for name, want in expected.items():
+        got = actual[name]
+        assert got["backend"] == want["backend"], name
+        assert got.get("n_traces") == want.get("n_traces"), name
+        assert got.get("n_axes") == want.get("n_axes"), name
